New: duff, for diff -u (and -r)
Handy wrapper for doing diff -u or diff -ur.

The -u format is so much better, it's pretty much always what's wanted.
But it isn't backwards-compatible with the default, and there may be
scripts which expect that (or want to use -c), so wrap it with a different
command name, duff.

This previously was a Bash function (and possibly an alias before that?).
Switching to a command so that it can be invoked inside sudo.

Nothing to do with the duff command from the duff package, which was only
added to Debian in 2012. I've been using this duff as a diff -u wrapper
for about a decade longer than that: my comment in the Bash function
linked to www.xray.mpe.mpg.de and had two spaces between sentences!

I couldn't work out precisely which email that comment had linked to
(http://www.xray.mpe.mpg.de/mailing-lists/perl5-porters/2002-08/msg01173.html,
to preserve history), but looking at that month's P5P emails, I've found a
couple of likely possibilities and linked those from the comments in this.