New: o
Handy short command for opening pretty much anything, which also doesn't
require me to remember which of the several mimeopen-like commands work.

Based on a command I found in a CVS directory. Except that since then
File::MimeInfo, and therefore mimeopen, has improved: it now opens
commands in the background itself, so exec-ing it with & is no longer
necessary — and as such, nor is a brief sleep before exiting, to ensure
that mimeopen's output appears before the prompt.

This now makes this wrapper very small indeed, but it's still worth it for
Huffman encoding and remembering — and for the possibility of switching to
a different command in future.