Fix: ffmpeg is ffmpeg again
This was an old version of this script I copied in, from the period when
the ffmpeg in Debian was usurped by a project that renamed it to avconv.
The original ffmpeg project has now regained the throne and the command is
called ffmpeg again.

I'm not going to bother with a compatibility layer that checks for both,
because it's now half a decade or something since avconv stopped being a
thing. We can all just use ffmpeg now and forget that avconv ever existed.