Enhancement: No ‘notify-send’ in notifications
Notifications were showing in Gnome as being sent by notify-send — which
is accurate, but unhelpful.

Instead, set the notification's application name to be the name of the
VPN.

This may have made things worse in other desktop environments, or with
other Gnome themes: I don't know if the application name is always
displayed; when it isn't, nothing is now showing the network.