New: ssh wrapper that invokes waypipe
There isn't a built-in way of Wayland-forwarding like there is
X-forwarding with SSH. Running waypipe ssh does that, so this is a wrapper
which automatically invokes it when in a Wayland environment.

Other options include setting up System D services on clients and all
servers with a little SSH config to connect them:
https://gist.github.com/dcode/912be128174c468a2dc28f263bdb7328
https://archive.is/wip/vcgmZ

Or using more SSH config and a helper script to run waypipe (a comment on
the above, with more suggests in further comments):
https://gist.github.com/dcode/912be128174c468a2dc28f263bdb7328?permalink_comment_id=5766855#gistcomment-5766855

Or maybe it would be possible to run the two halves of waypipe separately
with the ssh_config LocalCommand and RemoteCommand options?

But for now, this seems to work, so I'm trying it out as something that
involves the least moving parts.