Fix: Let git work
An ssh push to GitHub was failing, because GitHub (unsurprisingly) doesn't
have waypipe at the other end.

This is the disadvantage of this approach: it applies to all hosts, and,
unlike X11 forwarding, it doesn't fail back to just making the SSH
connection without forwarding.

I suspect this is too fragile and there need to be more exceptions than
just git. For instance, there's connecting to a new server for the first
time to install waypipe!